requirements.txt
discord.py==2.1.0
pynput==1.7.6
pillow==9.3.0
filesplit==4.0.1
pyaudio==0.2.12
sounddevice==0.4.5
scipy==1.9.3
pyautogui==0.9.53
pynacl==1.5.0
psutil==5.9.4
pypiwin32==223
pycryptodome==3.15.0
requests==2.28.1
browser_history==0.3.2
pygame
imageio
imageio[ffmpeg]
numpy
pynput
